// src/app/module/product/products.interface.ts
export type TVariant = {
  type: string;
  value: string;
};

export type TInventory = {
  quantity: number;
  inStock: boolean;
};

export type TProduct = {
  _id?: string;
  name: string;
  description: string;
  price: number;
  category: string;
  tags: string[];
  variants: Array<TVariant>;
  inventory: TInventory;
};